refactor: alert 인스턴스는 모델과 관련 없으므로 분리
